dom/animation/test/css-transitions/test_animation-starttime.html

<!doctype html>
<meta charset=utf-8>
<title>CSSTransition.startTime</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<div id="log"></div>
<script>

'use strict';

test(t => {
  const div = addDiv(t, {
    style: 'margin-left: 100px; transition: margin-left 100s linear 100s',
  });
  flushComputedStyle(div);
  div.style.marginLeft = '200px';
  const animation = div.getAnimations()[0];

  assert_equals(animation.startTime, null, 'startTime is unresolved');
}, 'The start time of a newly-created transition is unresolved');

test(t => {
  const div = addDiv(t, {
    style: 'margin-left: 100px; transition: margin-left 100s linear 100s',
  });
  flushComputedStyle(div);
  div.style.marginLeft = '200px';
  const animation = div.getAnimations()[0];

  const timelineTime = animation.timeline.currentTime;
  animation.startTime = timelineTime;

  assert_times_equal(
    animation.startTime,
    timelineTime,
    'Check setting of startTime actually works'
  );
}, 'The start time of a transition can be set');

promise_test(async t => {
  const div = addDiv(t, {
    style: 'margin-left: 100px; transition: margin-left 100s linear 100s',
  });
  flushComputedStyle(div);
  div.style.marginLeft = '200px';
  const animation = div.getAnimations()[0];

  await animation.ready;

  const timelineTime = animation.timeline.currentTime;
  const marginLeft = () => parseFloat(getComputedStyle(div).marginLeft);

  animation.startTime = timelineTime - 100 * MS_PER_SEC;
  assert_equals(marginLeft(), 100);

  animation.startTime = timelineTime - 150 * MS_PER_SEC;
  assert_equals(marginLeft(), 150);
}, 'The start time can be set to seek a transition');

promise_test(async t => {
  const div = addDiv(t, {
    style: 'margin-left: 100px; transition: margin-left 100s linear 100s',
  });
  const eventWatcher = new EventWatcher(t, div, [
    'transitionstart',
    'transitionend',
  ]);
  flushComputedStyle(div);
  div.style.marginLeft = '200px';
  const animation = div.getAnimations()[0];

  await animation.ready;

  const timelineTime = animation.timeline.currentTime;

  animation.startTime = timelineTime - 100 * MS_PER_SEC;
  await eventWatcher.wait_for('transitionstart');

  animation.startTime = timelineTime - 200 * MS_PER_SEC;
  await eventWatcher.wait_for('transitionend');
}, 'Seeking a transition using start time dispatches transition events');

</script>
